requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1714483155106/work
cachetools==5.5.0
certifi @ file:///croot/certifi_1734473278428/work/certifi
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
hyperframe==6.0.1
idna==3.10
ignr==2.2
Jinja2 @ file:///croot/jinja2_1730902924303/work
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-core==0.3.28
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.4
langsmith==0.2.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///work/perseverance-python-buildout/croot/mpmath_1698864994882/work
narwhals==1.20.1
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.58.1
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pillow @ file:///croot/pillow_1734430590456/work
portalocker==2.10.1
protobuf==5.29.2
pyarrow==18.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyPDF2==3.0.1
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
qdrant-client==1.12.2
referencing==0.35.1
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typer==0.15.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.3
websockets==14.1
wheel==0.44.0
XlsxWriter==3.2.0
